# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tekstil Otomasyonu API - OpenAPI 3.0",
        "license": {
            "name": "MIT License",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        }
    ],
...
    "paths": {
...
        "/musteriler": {
            "post": {
                "tags": [
                    "musteriler"
                ],
                "summary": "Müşteri oluşturur.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MusteriWithoutID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Müşteri Oluşturuldu.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Musteri"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Müşteri bilgileri eksik."
                    },
                    "500": {
                        "description": "Müşteri Kaydedilemedi."
                    }
                }
            },
            "get": {
                "tags": [
                    "musteriler"
                ],
                "summary": "Tüm müşterileri Getirir",
                "responses": {
                    "200": {
                        "description": "Müşteriler Getirildi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Musteri"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Müşteriler bulunamadı."
                    }
                }
            }
        },
        "/musteriler/{id}": {
            "get": {
                "tags": [
                    "musteriler"
                ],
                "summary": "ID bilgisine göre müşteri getirir.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "İstenen müşterinin ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "İstenen müşteri getirildi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Musteri"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Müşteri bulunamadı."
                    }
                }
            },
            "put": {
                "tags": [
                    "musteriler"
                ],
                "summary": "Müşteri günceller.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Güncellenecek müşterinin ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MusteriWithoutID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Müşteri Güncellendi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Musteri"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Müşteri Bulunamadı."
                    },
                    "500": {
                        "description": "Müşteri Güncellenemedi."
                    }
                }
            },
            "delete": {
                "tags": [
                    "musteriler"
                ],
                "summary": "Müşteri siler.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Silinmek istenen müşterinin ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "İstenen müşteri silindi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Musteri"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Müşteri bulunamadı."
                    },
                    "500": {
                        "description": "Müşteri silinemedi."
                    }
                }
            },
            "patch": {
                "tags": [
                    "musteriler"
                ],
                "summary": "Müşteri günceller.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Güncellenecek müşterinin ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MusteriWithoutID"
                            }
                        }
                    },
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Musteri"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Müşteri bilgileri eksik."
                    },
                    "500": {
                        "description": "Müşteri Kaydedilemedi."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Musteri": {
                "required": [
                    "musteri_id",
                    "ad",
                    "soyad",
                    "unvan",
                    "adres",
                    "sehir",
                    "ulke",
                    "telefon",
                    "sirket_ad"
                ],
                "type": "object",
                "properties": {
                    "musteri_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 12
                    },
                    "ad": {
                        "type": "string",
                        "example": "Taha"
                    },
                    "soyad": {
                        "type": "string",
                        "example": "Muslu"
                    },
                    "unvan": {
                        "type": "string",
                        "example": "Yönetici"
                    },
                    "adres": {
                        "type": "string",
                        "example": "Akasya mahallesi 123. sokak NO:12 D:3 Kadıköy"
                    },
                    "sehir": {
                        "type": "string",
                        "example": "İstanbul"
                    },
                    "ulke": {
                        "type": "string",
                        "example": "Türkiye"
                    },
                    "telefon": {
                        "type": "string",
                        "example": "532-123-4567"
                    },
                    "sirket_ad": {
                        "type": "string",
                        "example": "X Company"
                    }
                }
            },
            "MusteriWithoutID": {
                "required": [
                    "ad",
                    "soyad",
                    "unvan",
                    "adres",
                    "sehir",
                    "ulke",
                    "telefon",
                    "sirket_ad"
                ],
                "type": "object",
                "properties": {
                    "ad": {
                        "type": "string",
                        "example": "Taha"
                    },
                    "soyad": {
                        "type": "string",
                        "example": "Muslu"
                    },
                    "unvan": {
                        "type": "string",
                        "example": "Yönetici"
                    },
                    "adres": {
                        "type": "string",
                        "example": "Akasya mahallesi 123. sokak NO:12 D:3 Kadıköy"
                    },
                    "sehir": {
                        "type": "string",
                        "example": "İstanbul"
                    },
                    "ulke": {
                        "type": "string",
                        "example": "Türkiye"
                    },
                    "telefon": {
                        "type": "string",
                        "example": "532-123-4567"
                    },
                    "sirket_ad": {
                        "type": "string",
                        "example": "X Company"
                    }
                }
            },
...
        }
    }
}